feat: update user

+++ src/main/java/nus/iss/team3/backend/dataaccess/IUserAccountDataAccess.java
@@ -14,11 +14,11 @@ public interface IUserAccountDataAccess {
 
   boolean addUser(UserAccount userAccount);
 
-  boolean deleteUserById(Long id);
+  boolean deleteUserById(Integer id);
 
   boolean updateUser(UserAccount userAccount);
 
-  UserAccount getUserById(Long id);
+  UserAccount getUserById(Integer id);
 
   UserAccount getUserByEmail(String email);
